some extra clean up
