+++ src/hooks/index.tsx
@@ -1,4 +1,4 @@
 export * from './useGetNetworkConfig';
 export * from '../services/transactions/sendTransactions';
-export * from './useGetAccountInfo';
-export * from './useGetLoginInfo';
+export * from '../accountInfo/hooks/useGetAccountInfo';
+export * from '../loginInfo/hooks/useGetLoginInfo';
