+++ src/loginInfo/slice/reducers/login.ts
@@ -1,0 +1,10 @@
+import {LoginInfoStateType} from '../types';
+import {PayloadAction} from '@reduxjs/toolkit';
+import {LoginActionPayloadType} from '../../../redux/commonActions';
+
+export const login = (
+  state: LoginInfoStateType,
+  action: PayloadAction<LoginActionPayloadType>
+) => {
+  state.loginMethod = action.payload.loginMethod;
+};
