+++ src/loginInfo/slice/reducers/logout.ts
@@ -1,0 +1,3 @@
+import { initialState } from '../initialState';
+
+export const logout = () => initialState;
