+++ src/loginInfo/slice/reducers/setExtensionLogin.ts
@@ -1,0 +1,9 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+import { LoginInfoStateType, LoginInfoType } from '../types';
+
+export const setExtensionLogin = (
+  state: LoginInfoStateType,
+  action: PayloadAction<LoginInfoType>
+) => {
+  state.extensionLogin = action.payload;
+};
