+++ src/loginInfo/slice/reducers/setLedgerLogin.ts
@@ -1,0 +1,9 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+import { LedgerLoginType, LoginInfoStateType } from '../types';
+
+export const setLedgerLogin = (
+  state: LoginInfoStateType,
+  action: PayloadAction<LedgerLoginType>
+) => {
+  state.ledgerLogin = action.payload;
+};
