+++ src/loginInfo/slice/reducers/setTokenMethod.ts
@@ -1,0 +1,9 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+import { LoginInfoStateType, TokenLoginType } from '../types';
+
+export const setTokenLogin = (
+  state: LoginInfoStateType,
+  action: PayloadAction<TokenLoginType>
+) => {
+  state.tokenLogin = action.payload;
+};
