+++ src/loginInfo/slice/reducers/setWalletConnectLogin.ts
@@ -1,0 +1,9 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+import { LoginInfoStateType, WalletConnectLoginType } from '../types';
+
+export const setWalletConnectLogin = (
+  state: LoginInfoStateType,
+  action: PayloadAction<WalletConnectLoginType>
+) => {
+  state.walletConnectLogin = action.payload;
+};
