+++ src/loginInfo/slice/reducers/setWalletLogin.ts
@@ -1,0 +1,9 @@
+import { PayloadAction } from '@reduxjs/toolkit';
+import { LoginInfoStateType, LoginInfoType } from '../types';
+
+export const setWalletLogin = (
+  state: LoginInfoStateType,
+  { payload }: PayloadAction<LoginInfoType>
+) => {
+  state.walletLogin = payload;
+};
