+++ src/redux/slices/index.ts
@@ -1,6 +1,6 @@
-export * from './loginInfoSlice';
+export { actions as loginActions, loginInfoSlice } from '../../loginInfo/slice';
+export * from '../../accountInfo/accountInfoSlice';
 export * from './networkConfigSlice';
-export * from './accountInfoSlice';
 export * from './transactionsSlice';
 export * from './transactionsInfoSlice';
 export * from './modalsSlice';
