+++ src/services/login/types/index.ts
@@ -1,0 +1,16 @@
+import {
+  InitiateLoginFunctionType,
+  LoginHookGenericStateType
+} from '../../types';
+
+/** @internal */
+export type UseWebWalletLoginPropsType = {
+  callbackRoute: string;
+  token?: string;
+};
+
+/** @internal */
+export type UseWebWalletLoginReturnType = [
+  InitiateLoginFunctionType,
+  LoginHookGenericStateType
+];
